Automation with Power Query
In this release, took the data from a publicly shared folder by the data owner
https://onedrive.live.com/?authkey=%21AExIgOIi7vlj4mk&id=B09F9559F6A16B6C%2170480&cid=B09F9559F6A16B6C
and this data was actually sales reports of a fictional company. I used power query on 2017, 2018, 2019 sales report and transformed it into a more decent format and added a few columns helpful for analysis. Now, only what you need to do is to add the new sales report(2020) to the directory of pre-existing reports and power query will automatically perform all those steps for you.